pyproject.toml
[tool.poetry]
name = "ScriptCrew"
version = "0.1.0"
description = "Django-based multi-user system for managing content creation using CrewAI"
authors = ["ScriptCrew Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "4.2.11"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-filter = "^25.1"
django-cors-headers = "^4.7.0"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"
django-debug-toolbar = "^5.0.1"
crewai = "^0.105.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
